==> Churchill Cleaning/2 Churchill Flux and Met Cleaning.Rmd <==
---
title: "Churchill Cleaning"
output: pdf_document
date: "2022-12-08"
---

# To Do


Make sure all of the timestamps are loaded
Look at roll apply command

# Loading DF's

Clear Console and add packages


# Load 

```{r, warning=FALSE}
rm(list = ls())
Sys.setenv(TZ = "UTC")

library(data.table)
library(ggplot2)
library(cowplot)
library(openair)
library(plotrix)
library(signal)
library(svMisc)
library(zoo)
library(stringr)
library(plyr)
library(dplyr)


df = fread('C:/Users/dtrangmoe/Documents/Churchill/churchillFenBiomet_Aug2022-Jan2023.csv',na.strings = c('-9999','NA','NaN','NAN','-7999'))

h = fread('C:/Users/dtrangmoe/Documents/Churchill/Churchill_AllBiomet_until20230620.dat',skip=1 ,nrows = 0)
dat = fread('C:/Users/dtrangmoe/Documents/Churchill/Churchill_AllBiomet_until20230620.dat',skip=4, header = FALSE, na.strings=c('-9999','NA','NaN','NAN',"-7999"))

h2 = fread('C:/Users/dtrangmoe/Documents/Churchill/Churchill_AllBiomet_until20230712.dat',skip=1 ,nrows = 0)
dat2 = fread('C:/Users/dtrangmoe/Documents/Churchill/Churchill_AllBiomet_until20230712.dat',skip=4, header = FALSE, na.strings=c('-9999','NA','NaN','NAN',"-7999"))

h3 = fread('C:/Users/dtrangmoe/Documents/Churchill/Churchill_AllBiomet.dat',skip=1 ,nrows = 0)
dat3 = fread('C:/Users/dtrangmoe/Documents/Churchill/Churchill_AllBiomet.dat',skip=4, header = FALSE, na.strings=c('-9999','NA','NaN','NAN',"-7999"))

names(dat) = names(h)
names(dat2) = names(h2)
names(dat3) = names(h3)

met= dat
met2 = dat2
met3 = dat3

df = rbind.fill(df,met,met2,met3)

df = df[!duplicated(df$TIMESTAMP),]


write.csv(x = df,file = 'C:/Users/dtrangmoe/Documents/Churchill/Churchill_met_merged.csv',row.names = F,quote = F)

#check time stamps to ensure we have a complete series
TIMESTAMP = seq(from = min(df$TIMESTAMP), to = max(df$TIMESTAMP),by = 60*30,tz = "UTC")
tsdf = as.data.frame(TIMESTAMP)

df = merge(tsdf,df,by = "TIMESTAMP",all = T)


#na.strings not working for numbers for some reason, remove all here
df[df == -9999] = NA
df[df == -7999] = NA
summary(df)
```


Creates a variable to input a date range for the plots in the biomet section of
cleaning code. Used to easily change date range for different data sets. 
```{r, warning=FALSE}
biomet_dates <- c('2022-08-05', '2024-02-28')
```


# Air T
```{r, warning=FALSE}
ggplot(data = df)+
  geom_point(aes(TIMESTAMP,TA_2_1_1_1_1),col='black')+
  scale_y_continuous(#limits = c(-40,40),
                     expression('Air Temperature ('*Celsius*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))

#First we need to just convert from Kelvin to degrees C
df$TA_2_1_1_1_1 = ifelse(df$TA_2_1_1_1_1 > 200,df$TA_2_1_1_1_1 - 273.15,df$TA_2_1_1_1_1)

ggplot(data = df)+
  geom_point(aes(TIMESTAMP,TA_2_1_1_1_1),col='black')+
  scale_y_continuous(limits = c(-40,40),
                      expression('Air Temperature ('*Celsius*')'))
```


## Limits
```{r, warning=FALSE}
hist(df$TA_2_1_1_1_1,breaks=100)
plot(df$TIMESTAMP,df$TA_2_1_1_1_1,ylim = c(-50,50));abline(h=-50, col= 'red');abline(h=50, col='red')
TA_2_1_1_1_1.limits = replace(df$TA_2_1_1_1_1, df$TA_2_1_1_1_1 < -50 | df$TA_2_1_1_1_1 > 50,NA)
hist(TA_2_1_1_1_1.limits,breaks=100)
plot(TA_2_1_1_1_1.limits)
```


#Relative Humidity

```{r,warning=FALSE}
ggplot(data = df)+
  geom_point(aes(TIMESTAMP,RH_19_3_1_1_1),col='black')+
  scale_y_continuous(#limits = c(20,100),
                     expression('Relative Humidity ('*percent*')'))
#  scale_x_datetime(limits = as.POSIXct(biomet_dates))
```


## Limits
```{r,warning=FALSE}
RH_19_3_1_1_1.limits = df$RH_19_3_1_1_1

RH_19_3_1_1_1.limits = ifelse(df$RH_19_3_1_1_1 > 0 & df$TIMESTAMP <= '2022-9-23',NA, RH_19_3_1_1_1.limits)

df$RH_19_3_1_1_1.c = RH_19_3_1_1_1.limits

ggplot(data = df)+
  geom_point(aes(TIMESTAMP,RH_19_3_1_1_1.c),col='black')+
  scale_y_continuous(limits = c(5,100),
                     expression('Relative Humidity ('*percent*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))
```

# Barometric Pressure 

```{r,warning=FALSE}
ggplot(data = df)+
  geom_point(aes(TIMESTAMP,PA_4_2_1_1_1),col='black')+
  scale_y_continuous(#limits = c(90000,110000),
                     expression ('Barometric Pressure ('*kPa*')'))
#  scale_x_datetime(limits = as.POSIXct(biomet_dates))

df$PA_4_2_1_1_1 = ifelse(df$PA_4_2_1_1_1 > 97000,df$PA_4_2_1_1_1/1000,df$PA_4_2_1_1_1)

ggplot(data = df)+
  geom_point(aes(TIMESTAMP,PA_4_2_1_1_1),col='black')+
  scale_y_continuous(limits = c(0,200),
                     expression ('Barometric Pressure ('*kPa*')'))


df$PA_4_2_1_1_1 = ifelse(df$PA_4_2_1_1_1 > 150,df$PA_4_2_1_1_1/10,df$PA_4_2_1_1_1)

ggplot(data = df)+
  geom_point(aes(TIMESTAMP,PA_4_2_1_1_1),col='black')+
  scale_y_continuous(limits = c(0,140),
                     expression ('Barometric Pressure ('*kPa*')'))
```


```{r, warning=FALSE}
hist(df$PA_4_2_1_1_1,breaks=100)

ggplot(data = df)+
  geom_point(aes(TIMESTAMP,PA_4_2_1_1_1),col='black')+
  scale_y_continuous(limits = c(90,110),
                     expression ('Barometric Pressure ('*kPa*')'))

PA_4_2_1_1_1.limits = ifelse(df$PA_4_2_1_1_1 > 110 | df$PA_4_2_1_1_1 < 97,NA, df$PA_4_2_1_1_1)

ggplot(data = df)+
  geom_point(aes(TIMESTAMP,PA_4_2_1_1_1.limits),col='black')+
  geom_hline(yintercept = 106,col = 'red')+
  geom_hline(yintercept = 97,col = 'red')+
  scale_y_continuous(limits = c(90,110),
                     expression ('Barometric Pressure ('*kPa*')'))
```
## Rolling MAD
```{r, warning=FALSE}
f = PA_4_2_1_1_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data


ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,PA_4_2_1_1_1.limits,col='Pressure'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(0,110))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))
```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3 

#Add clean value to dataframe (_.c)

df$PA_4_2_1_1_1.c = ifelse(abs(PA_4_2_1_1_1.limits - mad_filt) > (N * roll.mad),NA,PA_4_2_1_1_1.limits)

#Compare cleaned vs. uncleaned

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,PA_4_2_1_1_1,col='Uncleaned'))+
  geom_point(aes(TIMESTAMP,PA_4_2_1_1_1.c,col='Cleaned'))+
  scale_y_continuous(limits = c(90,110))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))
```

```{r, warning=FALSE}
ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,PA_4_2_1_1_1.limits),col='red')+
  geom_point(aes(TIMESTAMP,PA_4_2_1_1_1.c),col='black')+
  scale_y_continuous(limits = c(80,115))+
scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))
```


#Soil Heat Flux
```{r, warning=FALSE}
ggplot(data = df)+
  geom_point(aes(TIMESTAMP,SHF_6_37_1_1_1,col='SHF 1'))+
  geom_point(aes(TIMESTAMP,SHF_6_37_2_1_1,col='SHF 2'))+
  scale_y_continuous(limits = c(-80,190),
                     expression ('Soil Heat Flux ('*W/m^2*')'))+
  scale_color_manual(values = c('red','black'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))
```

## SHF1
```{r, warning=FALSE}
hist(df$SHF_6_37_1_1_1,breaks=100)
plot(df$TIMESTAMP,df$SHF_6_37_1_1_1,ylim = c(-50,200));abline(h=-55, col= 'red');abline(h=190, col='red')
SHF_6_37_1_1_1.limits = replace(df$SHF_6_37_1_1_1, df$SHF_6_37_1_1_1 < -55 | df$SHF_6_37_1_1_1 > 190,NA)
hist(SHF_6_37_1_1_1.limits,breaks=100)
plot(SHF_6_37_1_1_1.limits)
df$SHF_6_37_1_1_1.c = SHF_6_37_1_1_1.limits
```

## SHF2
```{r, warning=FALSE}
hist(df$SHF_6_37_2_1_1,breaks=100)
plot(df$TIMESTAMP,df$SHF_6_37_2_1_1,ylim = c(-80,165));abline(h=-75, col= 'red');abline(h=150, col='red')
SHF_6_37_2_1_1.limits = replace(df$SHF_6_37_2_1_1, df$SHF_6_37_2_1_1 < -75 | df$SHF_6_37_2_1_1 > 150,NA)
hist(SHF_6_37_2_1_1.limits,breaks=100)
plot(SHF_6_37_2_1_1.limits)
df$SHF_6_37_2_1_1.c = SHF_6_37_2_1_1.limits
```

# Snow Depth
```{r, warning=FALSE}
ggplot(data = df)+
  geom_point(aes(TIMESTAMP,SD_10_99_1_1_1),col='black')+
  scale_y_continuous(limits = c(-1.5,.75),
                     expression('Snow Depth ('*m*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))
```


```{r, warning=FALSE}
hist(df$SD_10_99_1_1_1,breaks=100)
plot(df$TIMESTAMP,df$SD_10_99_1_1_1,ylim = c(-1.5,1));abline(h=0, col= 'red');abline(h=.8, col='red')
SD_10_99_1_1_1.limits=df$SD_10_99_1_1_1
SD_10_99_1_1_1.limits = ifelse(df$SD_10_99_1_1_1 < 1 & df$TIMESTAMP >= '2023-5-8' & df$TIMESTAMP <= '2023-10-28',NA,
                               ifelse(df$SD_10_99_1_1_1 < 0 | df$SD_10_99_1_1_1 > .8,NA,df$SD_10_99_1_1_1))
                               
ggplot(data = df)+
  geom_point(aes(TIMESTAMP,SD_10_99_1_1_1.limits),col='black')+
  scale_y_continuous(limits = c(-.25,.75),
                     expression('Snow Depth ('*m*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))
df$SD_10_99_1_1_1.c = SD_10_99_1_1_1.limits 
```

## Rolling MAD
```{r, warning=FALSE}
f = SD_10_99_1_1_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data

```

```{r, warning=FALSE}

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,SD_10_99_1_1_1.limits,col='Snow Depth'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-.1,.8))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))

```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 2

#Add clean value to dataframe (_.c)

df$SD_10_99_1_1_1.c = ifelse(abs(SD_10_99_1_1_1.limits - mad_filt) > (N * roll.mad),NA,SD_10_99_1_1_1.limits)

#Compare cleaned vs. uncleaned

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,SD_10_99_1_1_1,col='Uncleaned'))+
  geom_point(aes(TIMESTAMP,SD_10_99_1_1_1.c,col='Cleaned'))+
  scale_y_continuous(limits = c(-0.75,1))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))
```


```{r, warning=FALSE}
ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,SD_10_99_1_1_1.limits),col='red')+
  geom_point(aes(TIMESTAMP,SD_10_99_1_1_1.c),col='black')+
  scale_y_continuous(limits = c(-.25,.5))+
scale_x_datetime(limits = as.POSIXct(c('2024-1-1','2024-2-26'),format="%F"))

```


# PAR
```{r, warning=FALSE}
ggplot(data = df)+
  geom_point(aes(TIMESTAMP,PPFD_7_21_1_1_1),col='black')+
  scale_y_continuous(limits = c(-25,2000),
                     expression('PAR ('*mu*mol/m^2/s*')'))
```

```{r, warning=FALSE}
summary(df$PPFD_7_21_1_1_1)
hist(df$PPFD_7_21_1_1_1,breaks=100)
plot(df$TIMESTAMP,df$PPFD_7_21_1_1_1,ylim = c(0,1850));abline(h=1750, col= 'red');abline(h=0, col='red')
PPFD_7_21_1_1_1.limits = replace(df$PPFD_7_21_1_1_1, df$PPFD_7_21_1_1_1 < 0 | df$PPFD_7_21_1_1_1 > 1750,NA)
hist(PPFD_7_21_1_1_1.limits,breaks=100)
plot(PPFD_7_21_1_1_1.limits)
```


### Rolling SD 
(Changed to SD because MAD was cutting good data)
```{r, warning=FALSE}
f = PPFD_7_21_1_1_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 14*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = sd,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,PPFD_7_21_1_1_1.limits,col='PAR in'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-15,1800))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))

```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 4

#Add clean value to dataframe (_.c)

df$PPFD_7_21_1_1_1.c = ifelse(abs(PPFD_7_21_1_1_1.limits - mad_filt) > (N * roll.mad),NA,PPFD_7_21_1_1_1.limits)

#Compare cleaned vs. uncleaned

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,PPFD_7_21_1_1_1.limits,col='Uncleaned'))+
  geom_point(aes(TIMESTAMP,PPFD_7_21_1_1_1.c,col='Cleaned'))+
  scale_y_continuous(limits = c(0,2000))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))
```

```{r, warning=FALSE}
ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,PPFD_7_21_1_1_1.limits),col='red')+
  geom_point(aes(TIMESTAMP,PPFD_7_21_1_1_1.c),col='black')+
  scale_y_continuous(limits = c(-500,700))+
scale_x_datetime(limits = as.POSIXct(c('2024-01-01','2024-02-26'),format="%F"))
```


# PAR Reflected
```{r, warning=FALSE}
ggplot(data = df)+
  geom_point(aes(TIMESTAMP,PPFDR_7_23_1_1_1),col='black')+
  scale_y_continuous(limits = c(-25,1600),
                     expression('PAR Reflected ('*mu*mol/m^2/s*')'))
```

```{r, warning=FALSE}
summary(df$PPFDR_7_23_1_1_1)
hist(df$PPFDR_7_23_1_1_1,breaks=100)
plot(df$TIMESTAMP,df$PPFDR_7_23_1_1_1,ylim = c(0,5850));abline(h=1400, col= 'red');abline(h=-1, col='red')
PPFDR_7_23_1_1_1.limits=df$PPFDR_7_23_1_1_1
PPFDR_7_23_1_1_1.limits = ifelse(df$PPFDR_7_23_1_1_1 < -1 | df$PPFDR_7_23_1_1_1 > 600 & df$TIMESTAMP <= '2022-10-31',NA,df$PPFDR_7_23_1_1_1)
hist(PPFDR_7_23_1_1_1.limits,breaks=100)
plot(PPFDR_7_23_1_1_1.limits)
```


### Rolling SD
(Changed to SD because MAD was cutting good data)
```{r, warning=FALSE}
f = PPFDR_7_23_1_1_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = sd,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,PPFDR_7_23_1_1_1.limits,col='PAR in'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-5,1800))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))

```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 4

#Add clean value to dataframe (_.c)

df$PPFDR_7_23_1_1_1.c = ifelse(abs(PPFDR_7_23_1_1_1.limits - mad_filt) > (N * roll.mad),NA,PPFDR_7_23_1_1_1.limits)

#Compare cleaned vs. uncleaned

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,PPFDR_7_23_1_1_1,col='Uncleaned'))+
  geom_point(aes(TIMESTAMP,PPFDR_7_23_1_1_1.c,col='Cleaned'))+
  scale_y_continuous(limits = c(0,2000))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))
```

```{r, warning=FALSE}
ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*4,ymin = mad_filt-roll.mad*4,fill='4'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,PPFDR_7_23_1_1_1.limits),col='red')+
  geom_point(aes(TIMESTAMP,PPFDR_7_23_1_1_1.c),col='black')+
  scale_y_continuous(limits = c(-600,900))+
scale_x_datetime(limits = as.POSIXct(c('2024-01-01','2024-02-29'),format="%F"))
```

# Net Radiation 
```{r, warning=FALSE}
ggplot(data = df)+
  geom_point(aes(TIMESTAMP, RN_6_5_1_1_1),col='black')+
  scale_y_continuous(limits = c(-125,850),
                     expression('Net Radiation ('*W/m^2*')'))
```

```{r, warning=FALSE}
hist(df$RN_6_5_1_1_1,breaks=100)
plot(df$TIMESTAMP,df$RN_6_5_1_1_1,ylim = c(-400,850));abline(h=835,col= 'red');abline(h=-300, col='red')
RN_6_5_1_1_1.limits = replace(df$RN_6_5_1_1_1, df$RN_6_5_1_1_1 < -300 | df$RN_6_5_1_1_1 > 835,NA)
hist(RN_6_5_1_1_1.limits,breaks=100)
plot(df$TIMESTAMP,RN_6_5_1_1_1.limits)
```
## Rolling SD
```{r, warning=FALSE}
f = RN_6_5_1_1_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = sd,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data




#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3 

#Add clean value to dataframe (_.c)

df$RN_6_5_1_1_1.c = ifelse(abs(RN_6_5_1_1_1.limits - mad_filt) > (N * roll.mad),NA,RN_6_5_1_1_1.limits)

#Compare cleaned vs. uncleaned

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,RN_6_5_1_1_1.limits,col='Uncleaned'))+
  geom_point(aes(TIMESTAMP,RN_6_5_1_1_1.c,col='Cleaned'))+
  scale_y_continuous(limits = c(-250,800))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))
```

```{r, warning=FALSE}
ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,RN_6_5_1_1_1.limits),col='red')+
  geom_point(aes(TIMESTAMP,RN_6_5_1_1_1.c),col='black')+
  scale_y_continuous(limits = c(-500,600))+
scale_x_datetime(limits = as.POSIXct(c('2023-3-6','2023-3-20'),format="%F"))
```

# Shortwave in 
```{r, warning=FALSE}
ggplot(data = df)+
  geom_point(aes(TIMESTAMP,SWIN_6_10_1_1_1))+
  scale_y_continuous(#limits = c(-10,1000),
                     expression('Shortwave In ('*W/m^2*')'))
```

```{r, warning=FALSE}
hist(df$SWIN_6_10_1_1_1,breaks=100)
plot(df$TIMESTAMP,df$SWIN_6_10_1_1_1,ylim = c(0,1000));abline(h=900,col= 'red');abline(h=-1, col='red')
SWIN_6_10_1_1_1.limits <- df$SWIN_6_10_1_1_1
SWIN_6_10_1_1_1.limits[df$SWIN_6_10_1_1_1 < 0] <- 0
hist(SWIN_6_10_1_1_1.limits,breaks=100)
plot(df$TIMESTAMP,SWIN_6_10_1_1_1.limits)
```

## Rolling SD 
```{r, warning=FALSE}
f = SWIN_6_10_1_1_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = sd,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data


ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,SWIN_6_10_1_1_1.limits,col='Shortwave in'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-5,900))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))

```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 4

#Add clean value to dataframe (_.c)

df$SWIN_6_10_1_1_1.c = ifelse(abs(SWIN_6_10_1_1_1.limits - mad_filt) > (N * roll.mad),NA,SWIN_6_10_1_1_1.limits)

#Compare cleaned vs. uncleaned

plot(df$TIMESTAMP,df$SWIN_6_10_1_1_1,col='red',ylim = c(-1,925));points(df$TIMESTAMP,df$SWIN_6_10_1_1_1.c)
```

```{r, warning=FALSE} 
ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,SWIN_6_10_1_1_1.limits),col='red')+
  geom_point(aes(TIMESTAMP,SWIN_6_10_1_1_1.c),col='black')+
  scale_y_continuous(limits = c(-300,700))+
scale_x_datetime(limits = as.POSIXct(c('2022-11-6','2022-12-25'),format="%F"))
```


# Shortwave out 
```{r, warning=FALSE}
ggplot(data = df)+
  geom_point(aes(TIMESTAMP,SWOUT_6_11_1_1_1))+
  scale_y_continuous(limits = c(-10,750),
                     expression('Shortwave Out ('*W/m^2*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))
```

```{r, warning=FALSE}
hist(df$SWOUT_6_11_1_1_1,breaks=100)
plot(df$TIMESTAMP,df$SWOUT_6_11_1_1_1,ylim = c(0,800));abline(h=700,col= 'red');abline(h=-1, col='red')
SWOUT_6_11_1_1_1.limits <- df$SWOUT_6_11_1_1_1
SWOUT_6_11_1_1_1.limits[df$SWOUT_6_11_1_1_1 < 0] <- 0
hist(SWOUT_6_11_1_1_1.limits,breaks=100)
plot(df$TIMESTAMP,SWOUT_6_11_1_1_1.limits)
```


### Rolling SD
(Changed to SD because MAD was cutting good data)
```{r, warning=FALSE}
f = SWOUT_6_11_1_1_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = sd,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,SWOUT_6_11_1_1_1.limits,col='Shortwave out'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-5,700))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))


```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 4

#Add clean value to dataframe (_.c)

df$SWOUT_6_11_1_1_1.c = ifelse(abs(SWOUT_6_11_1_1_1.limits - mad_filt) > (N * roll.mad),NA,SWOUT_6_11_1_1_1.limits)

#Compare cleaned vs. uncleaned

plot(df$TIMESTAMP,df$SWOUT_6_11_1_1_1,col='red',ylim = c(0,800));points(df$TIMESTAMP,df$SWOUT_6_11_1_1_1.c)
```

```{r, warning=FALSE}
ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*4,ymin = mad_filt-roll.mad*4,fill='4'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,SWOUT_6_11_1_1_1.limits),col='red')+
  geom_point(aes(TIMESTAMP,SWOUT_6_11_1_1_1.c),col='black')+
  scale_y_continuous(limits = c(-250,300))+
scale_x_datetime(limits = as.POSIXct(c('2023-9-24','2023-11-2'),format="%F"))
```


# Albedo 
```{r, warning=FALSE}
albedo = df$SWOUT_6_11_1_1_1.c / df$SWIN_6_10_1_1_1.c

ggplot(data = df)+
  geom_point(aes(TIMESTAMP,albedo))+
  scale_y_continuous(limits = c(-1,2),
                     expression('Albedo'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))

albedo.limits = replace(albedo, albedo < .05 | albedo > 1,NA)


ggplot(data = df)+
  geom_point(aes(TIMESTAMP,albedo.limits))+
  scale_y_continuous(limits = c(-.5,1.5),
                     expression('Albedo'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))


```

```{r, warning=FALSE}
f = albedo.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,albedo.limits,col='Albedo'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(0,1))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','yellow','red'))
```


```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 2

#Add clean value to dataframe (_.c)

df$ALB_1_1_1.c = ifelse(abs(albedo.limits - mad_filt) > (N * roll.mad),NA,albedo.limits)

#Compare cleaned vs. uncleaned


ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,albedo.limits,col='Uncleaned'))+
  geom_point(aes(TIMESTAMP,df$ALB_1_1_1.c,col='Cleaned'))+
  scale_y_continuous(limits = c(0,1))+
  scale_x_datetime(limits = as.POSIXct(c('2023-01-01','2023-02-01'),format="%F"))+
  scale_color_manual(values=c('black','red'))


```

# Longwave In 
```{r, warning=FALSE}
ggplot(data = df)+
  geom_point(aes(TIMESTAMP,LWIN_6_14_1_1_1))+
  scale_y_continuous(limits = c(50,750),
                     expression('Longwave In ('*W/m^2*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))

```

```{r, warning=FALSE}
hist(df$LWIN_6_14_1_1_1,breaks=100)
plot(df$TIMESTAMP,df$LWIN_6_14_1_1_1,ylim = c(80,750));abline(h=450, col= 'red');abline(h=90, col='red')
LWIN_6_14_1_1_1.limits = replace(df$LWIN_6_14_1_1_1, df$LWIN_6_14_1_1_1 < 80 | df$LWIN_6_14_1_1_1 > 450,NA)
hist(LWIN_6_14_1_1_1.limits,breaks=100)

ggplot(data = df)+
  geom_point(aes(TIMESTAMP,LWIN_6_14_1_1_1.limits))+
  scale_y_continuous(limits = c(50,550),
                     expression('Longwave In ('*W/m^2*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))

df$LWIN_6_14_1_1_1.c = LWIN_6_14_1_1_1.limits
```


# Longwave out 
```{r, warning=FALSE}
ggplot(data = df)+
  geom_point(aes(TIMESTAMP,LWOUT_6_15_1_1_1))+
  scale_y_continuous(limits = c(150,750),
                     expression('Longwave Out ('*W/m^2*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))
```

```{r, warning=FALSE}
hist(df$LWOUT_6_15_1_1_1,breaks=100)
plot(df$TIMESTAMP,df$LWOUT_6_15_1_1_1,ylim = c(80,750));abline(h=550, col= 'red');abline(h=85, col='red')
LWOUT_6_15_1_1_1.limits = replace(df$LWOUT_6_15_1_1_1, df$LWOUT_6_15_1_1_1 < 85 | df$LWOUT_6_15_1_1_1 > 550,NA)
hist(LWOUT_6_15_1_1_1.limits,breaks=100)
plot(LWOUT_6_15_1_1_1.limits)
df$LWOUT_6_15_1_1_1.c = LWOUT_6_15_1_1_1.limits
```


# Convert Soil Temps in Kelvin to Celcius
```{r, warning=FALSE}

ts_columns <- names(df)[grepl("TS", names(df))]

for (col in ts_columns) {
  df[[col]] <- ifelse(df[[col]] > 150, df[[col]] - 273.15, df[[col]])
}

```

# Open Water Surface Therm

```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_3_1_1, col = 'Surface'))+
  geom_hline(yintercept = 0)+
  scale_y_continuous(limits = c(-15,25),
                     expression('Open Water Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))

``` 

```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP, TS_2_38_3_1_1, col = 'Surface'))+
  scale_y_continuous(limits = c(-15,10),
                     expression('Open Water Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2022-11-16",'2022-11-29')))

TS_2_38_3_1_1.limits = df$TS_2_38_3_1_1

TS_2_38_3_1_1.limits = ifelse(df$TS_2_38_3_1_1 < -5 & df$TIMESTAMP >= '2022-11-16' & df$TIMESTAMP <= '2022-11-29' ,NA, TS_2_38_3_1_1.limits)

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP, TS_2_38_3_1_1.limits, col = 'Surface'))+
  scale_y_continuous(expression('Open Water Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2023-02-09",'2023-06-12')))

TS_2_38_3_1_1.limits = ifelse(df$TS_2_38_3_1_1 <= -100 & df$TIMESTAMP >= '2023-2-1' & df$TIMESTAMP <= '2023-6-12' ,NA, TS_2_38_3_1_1.limits)

TS_2_38_3_1_1.limits = ifelse(df$TIMESTAMP >= '2023-04-20' & df$TIMESTAMP <= '2023-6-12' ,NA, TS_2_38_3_1_1.limits)

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP, TS_2_38_3_1_1.limits, col = 'Surface'))+
  scale_y_continuous(expression('Open Water Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2023-02-09",'2023-06-12')))

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP, TS_2_38_3_1_1.limits, col = 'Surface'))+
  scale_y_continuous(limits = c(-15,10),
                     expression('Open Water Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2022-11-16",'2022-11-29')))

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP, TS_2_38_3_1_1.limits, col = 'Surface'))+
  scale_x_datetime(limits = as.POSIXct(c('2023-04-04','2023-05-01')))

TS_2_38_3_1_1.limits = ifelse(df$TIMESTAMP >= '2023-04-07 1800' & df$TIMESTAMP <= '2023-04-24 0000' ,NA, TS_2_38_3_1_1.limits)
```

## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_3_1_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data


ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_1_1.limits,col='Open Water Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-15,30))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))
```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_3_1_1.c = ifelse(abs(TS_2_38_3_1_1.limits - mad_filt) > (N * roll.mad),NA,TS_2_38_3_1_1.limits)

#Compare cleaned vs. uncleaned

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_1_1.limits,col='Uncleaned'))+
  geom_point(aes(TIMESTAMP,df$TS_2_38_3_1_1.c,col='Cleaned'))+
  scale_y_continuous(limits = c(-15,30))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))

```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_1_1.limits),col='black')+
  geom_point(aes(TIMESTAMP,TS_2_38_3_1_1.c),col='red')+
  scale_y_continuous(limits = c(-15,30))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-20','2022-12-2'),format="%F"))
```


# Open Water 5 Cm depth Therm

```{r, warning=FALSE}

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_3_2_1, col = '10 cm'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))

TS_2_38_3_2_1.limits = df$TS_2_38_3_2_1

TS_2_38_3_2_1.limits = ifelse(df$TS_2_38_3_2_1 <= -25 & df$TIMESTAMP >= '2022-8-1' & df$TIMESTAMP <= '2023-1-30' ,NA, TS_2_38_3_2_1.limits)

df$TS_2_38_3_2_1.c = TS_2_38_3_2_1.limits

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_3_2_1.c, col = '10 cm'))+
  scale_x_datetime(limits = as.POSIXct(c("2022-8-2",'2023-11-1')))

plot(df$TS_2_38_3_2_1.c)
```
## Rolling MAD
NA here

# Open Water 10 cm Therm

```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_3_3_1, col = '10 cm'))+
  scale_y_continuous(limits = c(-15,20),
                     expression('Open Water Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))


ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP, TS_2_38_3_3_1, col = '10 cm'))+
  scale_y_continuous(limits = c(-5,5),
                     expression('Open Water Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2022-10-18",'2022-10-24')))+
  geom_vline(xintercept=as.POSIXct('2022-10-21 1200'))

TS_2_38_3_3_1.limits = df$TS_2_38_3_3_1

TS_2_38_3_3_1.limits = ifelse(df$TIMESTAMP >= '2022-10-21 2300' & df$TIMESTAMP <= '2022-10-22 1000' ,NA, TS_2_38_3_3_1.limits)

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP, TS_2_38_3_3_1.limits, col = '10 cm'))+
  scale_y_continuous(limits = c(-5,5),
                     expression('Open Water Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2022-10-18",'2022-10-24')))

```
## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_3_3_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data


ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_3_1.limits,col='Open Water Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-15,20))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))

```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_3_3_1.c = ifelse(abs(TS_2_38_3_3_1.limits - mad_filt) > (N * roll.mad),NA,TS_2_38_3_3_1.limits)

#Compare cleaned vs. uncleaned

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_3_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_3_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-15,20))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))



```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_3_1),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_3_3_1.c),col='black')+
  scale_y_continuous(limits = c(-10,15))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-15','2022-11-12'),format="%F"))
```

# Open Water 20 cm depth Therm

```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_3_4_1, col = '20 Cm'))+
  scale_y_continuous(limits = c(-12,12),
                     expression('Open Water Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))


TS_2_38_3_4_1.limits = replace(df$TS_2_38_3_4_1, df$TS_2_38_3_4_1 < 15 | df$TS_2_38_3_4_1.limits > -15,NA)

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP, TS_2_38_3_4_1, col = '20 cm'))+
  scale_y_continuous(limits = c(-5,5),
                     expression('Open Water Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2022-10-18",'2022-10-24')))

TS_2_38_3_4_1.limits = ifelse(df$TIMESTAMP >= '2022-10-21 2300' & df$TIMESTAMP <= '2022-10-22 1000' ,NA, TS_2_38_3_4_1.limits)

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP, TS_2_38_3_4_1.limits, col = '20 cm'))+
  scale_y_continuous(limits = c(-5,5),
                     expression('Open Water Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2022-10-18",'2022-10-24')))

```


## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_3_4_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data


ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_4_1.limits,col='Open Water 20 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-12,12))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))

```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_3_4_1.c = ifelse(abs(TS_2_38_3_4_1.limits - mad_filt) > (N * roll.mad),NA,TS_2_38_3_4_1.limits)

#Compare cleaned vs. uncleaned

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_4_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_4_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-12,12))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))

```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_4_1),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_3_4_1.c),col='black')+
  scale_y_continuous(limits = c(-5,5))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-20','2022-11-28'),format="%F"))
```

# Open Water 30 cm depth Therm

```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_3_5_1, col = '30 Cm'))+
  scale_y_continuous(limits = c(-10,8.5),
                     expression('Open Water Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))

TS_2_38_3_5_1.limits = df$TS_2_38_3_5_1

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP, TS_2_38_3_5_1))+
  scale_y_continuous(limits = c(-5,5),
                     expression('Open Water Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2022-10-18",'2022-10-24')))

TS_2_38_3_5_1.limits = ifelse(df$TIMESTAMP >= '2022-10-21 2300' & df$TIMESTAMP <= '2022-10-22 1000' ,NA, TS_2_38_3_5_1.limits)

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP, TS_2_38_3_5_1.limits))+
  scale_y_continuous(limits = c(-5,5),
                     expression('Open Water Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2022-10-18",'2022-10-24')))
```
## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_3_5_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data


ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_5_1.limits,col='Open Water 30 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-10,8.5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))
```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 2

#Add clean value to dataframe (_.c)

df$TS_2_38_3_5_1.c = ifelse(abs(TS_2_38_3_5_1.limits - mad_filt) > (N * roll.mad),NA,TS_2_38_3_5_1.limits)

#Compare cleaned vs. uncleaned

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_5_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_5_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-10,8.5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))

```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_5_1.limits),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_3_5_1.c),col='black')+
  scale_y_continuous(limits = c(-3,5))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-17','2022-11-10'),format="%F"))
```


# Open Water 40 cm depth Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_3_6_1, col = '40 Cm'))+
  scale_y_continuous(limits = c(-10,7.5),
                     expression('Open Water Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))

TS_2_38_3_6_1.limits = df$TS_2_38_3_6_1

TS_2_38_3_6_1.limits = ifelse(df$TIMESTAMP >= '2022-10-21 2300' & df$TIMESTAMP <= '2022-10-22 1000' ,NA, TS_2_38_3_6_1.limits)
```


## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_3_6_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data


ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_6_1.limits,col='Open Water 40 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-10,7.5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))

```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_3_6_1.c = ifelse(abs(TS_2_38_3_6_1.limits - mad_filt) > (N * roll.mad),NA,TS_2_38_3_6_1.limits)

#Compare cleaned vs. uncleaned

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_6_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_6_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-10,10))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))
```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_6_1.limits),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_3_6_1.c),col='black')+
  scale_y_continuous(limits = c(-10,7.5))+
scale_x_datetime(limits = as.POSIXct(c('2022-11-01','2022-11-10'),format="%F"))
```


# Open Water 50 cm depth Therm

```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_3_7_1, col = '50 Cm'))+
  scale_y_continuous(limits = c(-10,7),
                     expression('Open Water Therms Soil T ('*K*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))

TS_2_38_3_7_1.limits = df$TS_2_38_3_7_1

TS_2_38_3_7_1.limits = ifelse(df$TIMESTAMP >= '2022-10-21 2300' & df$TIMESTAMP <= '2022-10-22 1000' ,NA, TS_2_38_3_7_1.limits)
```


## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_3_7_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_7_1.limits,col='Open Water 50 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-10,7.5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))
```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_3_7_1.c = ifelse(abs(TS_2_38_3_7_1.limits - mad_filt) > (N * roll.mad),NA,TS_2_38_3_7_1.limits)

#Compare cleaned vs. uncleaned

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_7_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_7_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-10,10))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))
```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_7_1.limits),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_3_7_1.c),col='black')+
  scale_y_continuous(limits = c(-5,5))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-15','2022-11-10'),format="%F"))
```


# Open Water 60 cm depth Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_3_8_1, col = '60 Cm'))+
  scale_y_continuous(limits = c(-10,5),
                     expression('Open Water Therms Soil T ('*K*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))

TS_2_38_3_8_1.limits = df$TS_2_38_3_8_1

TS_2_38_3_8_1.limits = ifelse(df$TIMESTAMP >= '2022-10-21 2300' & df$TIMESTAMP <= '2022-10-22 1000' ,NA, TS_2_38_3_8_1.limits)
```
## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_3_8_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_8_1.limits,col='Open Water 60 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-10,5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))
```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_3_8_1.c = ifelse(abs(TS_2_38_3_8_1.limits - mad_filt) > (N * roll.mad),NA,TS_2_38_3_8_1.limits)

#Compare cleaned vs. uncleaned

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_8_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_8_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-10,10))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))
```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_8_1.limits),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_3_8_1.c),col='black')+
  scale_y_continuous(limits = c(-10,5))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-15','2022-11-18'),format="%F"))
```


# Open Water 70 cm depth Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_3_9_1, col = '70 Cm'))+
  scale_y_continuous(limits = c(-10,5),
                     expression('Open Water Therms Soil T ('*K*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))

TS_2_38_3_9_1.limits = df$TS_2_38_3_9_1

TS_2_38_3_9_1.limits = ifelse(df$TIMESTAMP >= '2022-10-21 2300' & df$TIMESTAMP <= '2022-10-22 1000' ,NA, TS_2_38_3_9_1.limits)
```
## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_3_9_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_9_1.limits,col='Open Water 70 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-10,5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))
```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_3_9_1.c = ifelse(abs(TS_2_38_3_9_1.limits - mad_filt) > (N * roll.mad),NA,TS_2_38_3_9_1.limits)

#Compare cleaned vs. uncleaned

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_8_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_8_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-10,10))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))
```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_9_1.limits),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_3_9_1.c),col='black')+
  scale_y_continuous(limits = c(-5,10))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-15','2022-11-20'),format="%F"))
```


# Open Water 80 cm depth Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_3_10_1, col = '80 Cm'))+
  scale_y_continuous(limits = c(-10,5),
                     expression('Open Water Therms Soil T ('*K*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))

TS_2_38_3_10_1.limits = df$TS_2_38_3_10_1

TS_2_38_3_10_1.limits = ifelse(df$TIMESTAMP >= '2022-10-21 2300' & df$TIMESTAMP <= '2022-10-22 1000' ,NA, TS_2_38_3_10_1.limits)
```
## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_3_10_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_10_1.limits,col='Open Water 80 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-8,3.5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))
```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 4

#Add clean value to dataframe (_.c)

df$TS_2_38_3_10_1.c = ifelse(abs(TS_2_38_3_10_1.limits - mad_filt) > (N * roll.mad),NA,TS_2_38_3_10_1.limits)

#Compare cleaned vs. uncleaned


ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_10_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_10_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-8,5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))

```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_10_1.limits),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_3_10_1.c),col='black')+
  scale_y_continuous(limits = c(-7.5,10))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-17','2022-11-3'),format="%F"))
```


# Open Water 90 cm depth Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_3_11_1, col = '90 Cm'))+
  scale_y_continuous(limits = c(-10,5),
                     expression('Open Water Therms Soil T ('*K*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))

TS_2_38_3_11_1.limits = df$TS_2_38_3_11_1

TS_2_38_3_11_1.limits = ifelse(df$TIMESTAMP >= '2022-10-21 2300' & df$TIMESTAMP <= '2022-10-22 1000' ,NA, TS_2_38_3_11_1.limits)
```

## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_3_11_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data


ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_11_1.limits,col='Open Water 80 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-8,3.5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))
```


```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_3_11_1.c = ifelse(abs(TS_2_38_3_11_1.limits - mad_filt) > (N * roll.mad),NA,TS_2_38_3_11_1.limits)

#Compare cleaned vs. uncleaned

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_11_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_11_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-8,5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))

```

```{r, warning=FALSE}
ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_11_1),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_3_11_1.c),col='black')+
  scale_y_continuous(limits = c(-7.5,10))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-17','2022-11-3'),format="%F"))
```


# Open Water 1 m depth Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_3_12_1, col = '1 m'))+
  scale_y_continuous(limits = c(-1000,10),
                     expression('Open Water Therms Soil T ('*K*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))

TS_2_38_3_12_1.limits = df$TS_2_38_3_12_1

TS_2_38_3_12_1.limits = ifelse(df$TS_2_38_3_12_1 <= -40 & df$TIMESTAMP >= '2022-10-15' & df$TIMESTAMP <= '2022-12-15' ,NA, TS_2_38_3_12_1.limits)

TS_2_38_3_12_1.limits = ifelse(df$TIMESTAMP >= '2022-10-21 2300' & df$TIMESTAMP <= '2022-10-22 1000' ,NA, TS_2_38_3_12_1.limits)
```


## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_3_12_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data


ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_12_1.limits,col='Open Water 1 m Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-25,3.5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))
```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 1.5

#Add clean value to dataframe (_.c)

df$TS_2_38_3_12_1.c = ifelse(abs(TS_2_38_3_12_1.limits - mad_filt) > (N * roll.mad),NA,TS_2_38_3_12_1.limits)

#Compare cleaned vs. uncleaned

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_3_12_1.limits,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_12_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-8,5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))

```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_3_12_1),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_3_12_1.c),col='black')+
  scale_y_continuous(limits = c(-15,2))+
scale_x_datetime(limits = as.POSIXct(c('2023-4-4','2023-7-3'),format="%F"))
```


# All open water therms
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+
  geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP,TS_2_38_3_1_1.c,color='Surface'))+
  geom_line(aes(TIMESTAMP,TS_2_38_3_2_1.c,color='5 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_3_3_1.c,color='10 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_3_4_1.c,color='20 cm'))+  
  geom_line(aes(TIMESTAMP,TS_2_38_3_5_1.c,color='30 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_3_6_1.c,color='40 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_3_7_1.c,color='50 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_3_8_1.c,color='60 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_3_9_1.c,color='70 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_3_10_1.c,color='80 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_3_11_1.c,color='90 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_3_12_1.c,color='1 m')) 
```


# Hummock Surface Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_1_1, col = 'Surface'))+
  scale_y_continuous(limits = c(-15,35),
                     expression('Hummock Surface Therm Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))
TS_2_38_4_1_1 = df$TS_2_38_4_1_1
```
## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_4_1_1 # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_1_1,col='Hummock Surface therm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-15,35))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','black','red'))
```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_4_1_1.c = ifelse(abs(TS_2_38_4_1_1 - mad_filt) > (N * roll.mad),NA,TS_2_38_4_1_1)

#Compare cleaned vs. uncleaned

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_1_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_1_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-15,35))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))

```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_1_1),col='black')+
  geom_point(aes(TIMESTAMP,TS_2_38_4_1_1.c),col='red')+
  scale_y_continuous(limits = c(-10,10))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-17','2022-11-1'),format="%F"))
```
# Hummock 5 cm Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_2_1, col = 'Surface'))+
  scale_y_continuous(limits = c(-15,30),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))
TS_2_38_4_2_1 = df$TS_2_38_4_2_1
```
## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_4_2_1 # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data


ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_2_1,col='Open Water 80 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-15,30))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))

```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_4_2_1.c = ifelse(abs(TS_2_38_4_2_1 - mad_filt) > (N * roll.mad),NA,TS_2_38_4_2_1)

#Compare cleaned vs. uncleaned

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_2_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_2_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-15,30))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))

```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_2_1),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_4_2_1.c),col='black')+
  scale_y_continuous(limits = c(-15,25))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-17','2022-11-1'),format="%F"))
```

# Hummock 10 cm Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_3_1, col = 'Surface'))+
  scale_y_continuous(limits = c(-15,25),
                     expression('Hummock 10 cm Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))
TS_2_38_4_3_1 = df$TS_2_38_4_3_1
```


## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_4_3_1 # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data


ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_3_1,col='Open Water 80 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-15,22))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))


```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_4_3_1.c = ifelse(abs(TS_2_38_4_3_1 - mad_filt) > (N * roll.mad),NA,TS_2_38_4_3_1)

#Compare cleaned vs. uncleaned

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_3_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_3_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-15,25))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))
```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_3_1),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_4_3_1.c),col='black')+
  scale_y_continuous(limits = c(-15,20))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-17','2022-11-1'),format="%F"))
```

# Hummock 20 cm Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_4_1, col = 'Surface'))+
  scale_y_continuous(limits = c(-15,15),
                     expression('Hummock Therms 20 cm Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))
TS_2_38_4_4_1 = df$TS_2_38_4_4_1
```


## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_4_4_1 # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_4_1,col='Open Water 80 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-15,22))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','red','black'))

```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_4_4_1.c = ifelse(abs(TS_2_38_4_4_1 - mad_filt) > (N * roll.mad),NA,TS_2_38_4_4_1)

#Compare cleaned vs. uncleaned

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_4_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_4_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-15,15))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))

```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_4_1),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_4_4_1.c),col='black')+
  scale_y_continuous(limits = c(-15,15))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-17','2022-11-1'),format="%F"))
```


# Hummock 30 cm Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_5_1, col = 'Surface'))+
  scale_y_continuous(limits = c(-12,10),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))
TS_2_38_4_5_1 = df$TS_2_38_4_5_1


```


## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_4_5_1 # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_5_1,col='Hummock 30 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-15,22))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','black','red'))

```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_4_5_1.c = ifelse(abs(TS_2_38_4_5_1 - mad_filt) > (N * roll.mad),NA,TS_2_38_4_5_1)

#Compare cleaned vs. uncleaned

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_5_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_5_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-15,15))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))

```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_5_1),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_4_5_1.c),col='black')+
  scale_y_continuous(limits = c(-5,5))+
scale_x_datetime(limits = as.POSIXct(c('2023-5-17','2023-7-1'),format="%F"))
```
# Hummock 40 cm Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_6_1, col = 'Surface'))+
  scale_y_continuous(limits = c(-15,15),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))
TS_2_38_4_6_1 = df$TS_2_38_4_6_1
```
## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_4_6_1 # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_6_1,col='Hummock 40 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-15,10))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','black','red'))
```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_4_6_1.c = ifelse(abs(TS_2_38_4_6_1 - mad_filt) > (N * roll.mad),NA,TS_2_38_4_6_1)

#Compare cleaned vs. uncleaned

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_6_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_6_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-15,15))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))


```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_6_1),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_4_6_1.c),col='black')+
  scale_y_continuous(limits = c(-15,15))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-17','2022-11-1'),format="%F"))
```
# Hummock 50 cm Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_7_1, col = 'Surface'))+
  scale_y_continuous(limits = c(-10,7.5),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))
TS_2_38_4_7_1 = df$TS_2_38_4_7_1
```
## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_4_7_1 # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data


ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_7_1,col='Hummock 50 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-10,7.5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','black','red'))
```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_4_7_1.c = ifelse(abs(TS_2_38_4_7_1 - mad_filt) > (N * roll.mad),NA,TS_2_38_4_7_1)

#Compare cleaned vs. uncleaned
ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_7_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_7_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-15,15))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))
```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_7_1),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_4_7_1.c),col='black')+
  scale_y_continuous(limits = c(-5,10))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-17','2022-11-1'),format="%F"))
```
# Hummock 60 cm Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_8_1, col = 'Surface'))+
  scale_y_continuous(limits = c(-10,10),
                     expression('Hummock Therms Soil T ('*K*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))
TS_2_38_4_8_1 = df$TS_2_38_4_8_1
```

## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_4_8_1 # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data


ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_8_1,col='Hummock 50 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-10,7.5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','black','red'))
```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_4_8_1.c = ifelse(abs(TS_2_38_4_8_1 - mad_filt) > (N * roll.mad),NA,TS_2_38_4_8_1)

#Compare cleaned vs. uncleaned
ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_8_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_8_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-15,15))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))
```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_8_1),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_4_8_1.c),col='black')+
  scale_y_continuous(limits = c(-10,10))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-17','2022-11-1'),format="%F"))
```
# Hummock 70 cm Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_9_1, col = 'Surface'))+
  scale_y_continuous(limits = c(-10,5),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))
TS_2_38_4_9_1 = df$TS_2_38_4_9_1
```

## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_4_9_1 # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_9_1,col='Hummock 50 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-10,7.5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','black','red'))
```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_4_9_1.c = ifelse(abs(TS_2_38_4_9_1 - mad_filt) > (N * roll.mad),NA,TS_2_38_4_9_1)

#Compare cleaned vs. uncleaned
ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_9_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_9_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-10,5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))
```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_9_1),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_4_9_1.c),col='black')+
  scale_y_continuous(limits = c(-5,5))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-17','2022-11-1'),format="%F"))
```
# Hummock 80 cm Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_10_1, col = 'Surface'))+
  scale_y_continuous(limits = c(-10,5),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))
TS_2_38_4_10_1 = df$TS_2_38_4_10_1
```


## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_4_10_1 # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_10_1,col='Hummock 50 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-10,7.5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','black','red'))
```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_4_10_1.c = ifelse(abs(TS_2_38_4_10_1 - mad_filt) > (N * roll.mad),NA,TS_2_38_4_10_1)

#Compare cleaned vs. uncleaned
ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_10_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_10_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-10,5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))

```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_10_1),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_4_10_1.c),col='black')+
  scale_y_continuous(limits = c(-5,5))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-14','2022-11-15'),format="%F"))
```
# Hummock 90 cm Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_11_1, col = '90cm'))+
  scale_y_continuous(limits = c(-6,5),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2022-9-6",'2023-11-1')))

TS_2_38_4_11_1.limits = df$TS_2_38_4_11_1
```
## Limits
```{r, warning=FALSE}

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_11_1, col = '90cm'))+
  scale_y_continuous(limits = c(-6,3),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2022-9-6",'2022-11-20')))

TS_2_38_4_11_1.limits = ifelse(df$TS_2_38_4_11_1 <= -1 & df$TIMESTAMP >= '2022-10-15' & df$TIMESTAMP <= '2022-11-15' ,NA, TS_2_38_4_11_1.limits)

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_11_1.limits, col = '90cm'))+
  scale_y_continuous(limits = c(-6,3),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2022-9-6",'2022-11-20')))

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_11_1, col = '90cm'))+
  scale_y_continuous(limits = c(-6,3),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2022-11-20",'2023-1-20')))


TS_2_38_4_11_1.limits = ifelse(df$TS_2_38_4_11_1 <= -2 & df$TIMESTAMP >= '2022-11-20' & df$TIMESTAMP <= '2023-1-20' ,NA, TS_2_38_4_11_1.limits)


TS_2_38_4_11_1.limits = ifelse(df$TS_2_38_4_11_1 <= -1 & df$TIMESTAMP >= '2022-11-20' & df$TIMESTAMP <= '2022-12-20' ,NA, TS_2_38_4_11_1.limits)

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_11_1.limits, col = '90cm'))+
  scale_y_continuous(limits = c(-6,3),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2022-11-20",'2023-1-20')))

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_11_1, col = '90cm'))+
  scale_y_continuous(limits = c(-6,0),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2023-1-20",'2023-2-1')))

TS_2_38_4_11_1.limits = ifelse(df$TS_2_38_4_11_1 <= -1 & df$TIMESTAMP >= '2023-1-20' & df$TIMESTAMP <= '2023-2-1' ,NA, TS_2_38_4_11_1.limits)


ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_11_1.limits, col = '90cm'))+
  scale_y_continuous(limits = c(-6,0),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2023-1-20",'2023-2-1')))

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_11_1, col = '90cm'))+
  scale_y_continuous(limits = c(-6,0),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2023-2-1",'2023-2-14')))


TS_2_38_4_11_1.limits = ifelse(df$TS_2_38_4_11_1 <= -4 & df$TIMESTAMP >= '2023-2-1' & df$TIMESTAMP <= '2023-2-14' ,NA, TS_2_38_4_11_1.limits)

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_11_1.limits, col = '90cm'))+
  scale_y_continuous(limits = c(-6,0),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2023-2-1",'2023-2-14')))

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_11_1, col = '90cm'))+
  scale_y_continuous(limits = c(-6,0),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2023-2-14",'2023-2-20')))

TS_2_38_4_11_1.limits = ifelse(df$TS_2_38_4_11_1 <= -4.25 & df$TIMESTAMP >= '2023-2-14' & df$TIMESTAMP <= '2023-2-20' ,NA, TS_2_38_4_11_1.limits)

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_11_1.limits, col = '90cm'))+
  scale_y_continuous(limits = c(-6,0),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2023-2-14",'2023-2-20')))

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_11_1, col = '90cm'))+
  scale_y_continuous(limits = c(-10,0),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2023-2-20",'2023-2-28')))

TS_2_38_4_11_1.limits = ifelse(df$TS_2_38_4_11_1 <= -5 & df$TIMESTAMP >= '2023-2-20' & df$TIMESTAMP <= '2023-2-23' ,NA, TS_2_38_4_11_1.limits)

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_11_1.limits, col = '90cm'))+
  scale_y_continuous(limits = c(-10,0),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2023-2-20",'2023-2-28')))

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_11_1.limits, col = '90cm'))+
  scale_y_continuous(limits = c(-10,0),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2023-5-10",'2023-5-24')))

TS_2_38_4_11_1.limits = ifelse(df$TS_2_38_4_11_1 <= -2.5 & df$TIMESTAMP >= '2023-5-14' & df$TIMESTAMP <= '2023-6-23' ,NA, TS_2_38_4_11_1.limits)
```

## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_4_11_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_11_1.limits,col='Hummock 90 cm Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-10,7.5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','black','red'))
```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_4_11_1.c = ifelse(abs(TS_2_38_4_11_1.limits - mad_filt) > (N * roll.mad),NA,TS_2_38_4_11_1.limits)

#Compare cleaned vs. uncleaned

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_11_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_11_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-10,5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))

```

```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_11_1),col='red')+
  geom_point(aes(TIMESTAMP,TS_2_38_4_11_1.c),col='black')+
  scale_y_continuous(limits = c(-5,5))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-27','2022-11-30'),format="%F"))
```
# Hummock 1 m Therm
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_12_1, col = 'Surface'))+
  scale_y_continuous(limits = c(-10,10),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))

TS_2_38_4_12_1.limits = df$TS_2_38_4_12_1

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_12_1, col = 'Surface'))+
  scale_y_continuous(limits = c(-2,-.4),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2023-6-6",'2023-7-15')))

TS_2_38_4_12_1.limits = ifelse(df$TS_2_38_4_12_1 >= -.80 & df$TIMESTAMP >= '2023-06-15' & df$TIMESTAMP <= '2023-07-01' ,NA, TS_2_38_4_12_1.limits)

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_12_1.limits, col = 'Surface'))+
  scale_y_continuous(limits = c(-2,-.4),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2023-6-29",'2023-7-5')))

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_12_1, col = 'Surface'))+
  scale_y_continuous(limits = c(-2,-.4),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2023-7-2",'2023-8-10')))

TS_2_38_4_12_1.limits = ifelse(df$TS_2_38_4_12_1 <= -.90 & df$TIMESTAMP >= '2023-07-18' & df$TIMESTAMP <= '2023-08-10' ,NA, TS_2_38_4_12_1.limits)

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_12_1.limits, col = 'Surface'))+
  scale_y_continuous(limits = c(-2,-.4),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2023-7-2",'2023-8-10')))

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_12_1.limits, col = 'Surface'))+
  scale_y_continuous(limits = c(-2,-.4),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2023-6-30",'2023-7-20')))

TS_2_38_4_12_1.limits = ifelse(df$TS_2_38_4_12_1 < -1.1 & df$TIMESTAMP >= '2023-07-2' & df$TIMESTAMP <= '2023-07-18' ,NA, TS_2_38_4_12_1.limits)

ggplot(data = df)+ theme_bw()+ geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP, TS_2_38_4_12_1.limits, col = 'Surface'))+
  scale_y_continuous(limits = c(-2,-.4),
                     expression('Hummock Therms Soil T ('*C*')'))+
  scale_x_datetime(limits = as.POSIXct(c("2023-7-2",'2023-7-18')))
```

## Rolling MAD 
```{r, warning=FALSE}
f = TS_2_38_4_12_1.limits # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
mad_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
mad_filt = f                          # make a copy of the full fluxes again
mad_filt[ind] = mad_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = mad_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(mad_filt)] = NA  #make NA values where NAs exist in the real data


ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_12_1,col='Hummock 1 m Soil T'))+
  geom_line(aes(TIMESTAMP,mad_filt,col='Filter'))+
  geom_line(aes(TIMESTAMP,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-10,7.5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('yellow','black','red'))
```

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$TS_2_38_4_12_1.c = ifelse(abs(TS_2_38_4_12_1.limits - mad_filt) > (N * roll.mad),NA,TS_2_38_4_12_1.limits)

#Compare cleaned vs. uncleaned

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,TS_2_38_4_12_1,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_12_1.c,col='cleaned'))+
  scale_y_continuous(limits=c(-10,5))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))
```


```{r, warning=FALSE}

ggplot(data = df)+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*6,ymin = mad_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*3,ymin = mad_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*2,ymin = mad_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = TIMESTAMP,ymax = mad_filt+roll.mad*1,ymin = mad_filt-roll.mad*1,fill='1'))+
  geom_point(aes(TIMESTAMP,TS_2_38_4_12_1),col='black')+
  geom_point(aes(TIMESTAMP,TS_2_38_4_12_1.c),col='red')+
  scale_y_continuous(limits = c(-5,3))+
scale_x_datetime(limits = as.POSIXct(c('2022-10-27','2022-11-30'),format="%F"))
```
# All Hummock Therms
```{r, warning=FALSE}
ggplot(data = df)+ theme_bw()+
  geom_hline(yintercept = 0)+
  geom_line(aes(TIMESTAMP,TS_2_38_4_1_1.c,color='Surface'))+
  geom_line(aes(TIMESTAMP,TS_2_38_4_2_1.c,color='5 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_4_3_1.c,color='10 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_4_4_1.c,color='20 cm'))+  
  geom_line(aes(TIMESTAMP,TS_2_38_4_5_1.c,color='30 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_4_6_1.c,color='40 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_4_7_1.c,color='50 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_4_8_1.c,color='60 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_4_9_1.c,color='70 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_4_10_1.c,color='80 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_4_11_1.c,color='90 cm'))+
  geom_line(aes(TIMESTAMP,TS_2_38_4_12_1.c,color='1 m')) 
  

```

# Volumetric Water Content
```{r, warning=FALSE}
ggplot(data = df)+
  geom_point(aes(TIMESTAMP, SWC_12_36_1_1_1*100,col='Subwater Peat'))+
  geom_point(aes(TIMESTAMP, SWC_12_36_2_1_1*100,col='Surface Hummock'))+
  geom_point(aes(TIMESTAMP, SWC_12_36_2_2_1*100,col='10 cm Hummock'))+
  scale_y_continuous(limits = c(0,100),
                     expression('% VWC'))+
  scale_x_datetime(limits = as.POSIXct(biomet_dates))
```

# Save

```{r, warning=FALSE}
write.csv(x = df,file = 'C:/Users/dtrangmoe/Documents/Churchill/Churchill_met_clean.csv',row.names = F,quote = F)
```


# Flux Cleaning


```{r, warning=FALSE}
rm(list = ls())

library(data.table)
library(ggplot2)
library(cowplot)
library(openair)
library(plotrix)
library(signal)
library(svMisc)
library(zoo)
library(stringr)
library(plyr)
library(dplyr)
```

Load in eddypro output files

```{r warning=FALSE}

met = fread('C:/Users/dtrangmoe/Documents/Churchill/Churchill_met_clean.csv')

df = fread('C:/Users/dtrangmoe/Documents/Churchill/Churchill_fluxes_merged.csv')

df$TIMESTAMP = df$ts

df= merge(df,met,by='TIMESTAMP',all = T)

#check time stamps to ensure we have a complete series
TIMESTAMP = seq(from = min(df$TIMESTAMP), to = max(df$TIMESTAMP),by = 60*30,tz = "UTC")
tsdf = as.data.frame(TIMESTAMP)

df = merge(tsdf,df,by = "TIMESTAMP",all = T)

```


Creates a variable to input a date range for the plots in the flux section of
cleaning code. Used to easily change date range for different data sets. 
```{r, warning=FALSE}
flux_dates <- c('2022-08-05', '2024-01-15')
```
# Initial Clean

## Energy Balance

combine energy fluxes by met and flux components

```{r}
df$met = (df$LWIN_6_14_1_1_1.c + df$SWIN_6_10_1_1_1.c) - (df$LWOUT_6_15_1_1_1.c + df$SWOUT_6_11_1_1_1.c)
df$flux = df$H + df$LE

plot(df$met, df$flux)

#add better plot here! with 1:1 line and better axis ratio

```

plot all solar radiation

```{r, warning=FALSE}
plot(df$ts,df$SWIN_6_10_1_1_1.c)
plot(df$ts,df$SWOUT_6_11_1_1_1.c)
plot(df$ts,df$LWIN_6_14_1_1_1.c,ylim = c(0,800))
plot(df$ts,df$LWOUT_6_15_1_1_1.c,ylim = c(0,800))
```

create and plot mismatch variable between met and flux

```{r, warning=FALSE}
#df$mismatch = df$flux/df$met

#mismatch.limits = ifelse(df$mismatch > 10 & df$mismatch < -10, NA,mismatch.limits)

#mismatch.limits = replace(df$mismatch, df$mismatch < -10 | df$mismatch > 10,NA)


#plot(df$ts,mismatch.limits,)
```

plot energy balance

```{r, echo=FALSE}

#mismatch = df$mismatch

#ggplot(data = df)+ theme_bw()+ geom_abline(intercept = 1,color='blue')+
  #geom_point(aes(met,flux,color=mismatch.limits))+
  #scale_x_continuous(limits = c(-200,600))+
  #scale_y_continuous(limits = c(-200,600))+
  #scale_color_viridis()
    
#plot(df$met,df$flux, abline(0, 1, col= 'blue'))
```

NEED TO REASSESS OR GET RID OF THIS CHUNK ---- removes where the energy balance is off by more than a factor of 10, not sure I like this, mostly removes small fluxes where the factor would have a larger relative mismatch

create new variable _.energy

```{r, warning=FALSE}
#Tau.energy  = replace(df$Tau ,abs(df$mismatch)>10,NA)
#H.energy    = replace(df$H   ,abs(df$mismatch)>10,NA)
#LE.energy   = replace(df$LE  ,abs(df$mismatch)>10,NA)
#co2_flux.energy   = replace(df$co2_flux  ,abs(df$mismatch)>10,NA)
#ch4_flux.energy = replace(df$ch4_flux,abs(df$mismatch)>10,NA)
```

## EP QC

filter by the eddypro flags and create new variable _.qc

```{r, warning=FALSE}
Tau.qc      = replace(df$Tau,df$qc_Tau == 2,NA)
co2_flux.qc = replace(df$co2_flux,df$qc_co2_flux == 2,NA)
ch4_flux.qc = replace(df$ch4_flux,df$qc_ch4_flux == 2,NA)
LE.qc       = replace(df$LE,df$qc_LE == 2,NA)
H.qc        = replace(df$H,df$qc_H == 2,NA)
```

## Clear Outliers

look for absolute outliers and manually set some absolute limits to remove crazy data

### Tau

Plot Tau Histogram

```{r, warning=FALSE}
hist(df$Tau,breaks=100)
```

Cut the crazy data and create new variable (Tau.limits)

```{r, warning=FALSE}

Tau.limits = df$Tau
Tau.limits = replace(df$Tau,df$Tau < -1 | df$Tau > .2,NA)

plot(Tau.limits)

```

### NEE

Plot NEE Histogram and what we might clean

```{r, warning=FALSE}
hist(df$co2_flux,breaks=100)

#for 2022
plot(df$ts,df$co2_flux,ylim = c(-25,25));abline(h=10, col= 'red');abline(h=-14, col='red')

# for 2023 and on
plot(df$ts,df$co2_flux,ylim = c(-25,25));abline(h=7, col= 'red');abline(h=-10, col='red')
```

Clean based on graphs, create new variable (co2_flux.limits), and see cleaned product

```{r, warning=FALSE}
co2_flux.limits = replace(df$co2_flux,(df$co2_flux > 7 | df$co2_flux < -10) & df$TIMESTAMP >= '2023-01-01' | (df$co2_flux > 10 | df$co2_flux < -14),NA)


plot(df$ts,co2_flux.limits,ylim = c(-25,25));abline(h=7, col= 'red');abline(h=-10, col='red')

hist(co2_flux.limits,breaks=100)
```

### CH4

Plot CH4 Histogram and what we might clean

```{r, warning=FALSE}
hist(df$ch4_flux,breaks=100)

plot(df$ts,df$ch4_flux,ylim = c(-.3,.35));abline(h=.3, col= 'red');abline(h=-.20, col='red'); abline(h=0,col= 'black')
```

Clean based on graphs, create new variable (ch4_flux.limits), and see cleaned product

```{r, warning=FALSE}


CH4MOLE = ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(x=ts,y=ch4_mole_fraction))+
  geom_vline(xintercept = as.POSIXct('2023-9-14 16:30'),col = 'red')+
  geom_vline(xintercept = as.POSIXct('2023-9-21 11:00'),col = 'red')+
  scale_y_continuous(limits = c(0,4))+
  scale_x_datetime(limits = as.POSIXct(c('2023-9-5','2023-10-2'),format="%F"))
CH4MOLE

ch4_flux.limits = df$ch4_flux

ch4_flux.limits = ifelse(df$ch4_flux < 3 & df$ts <= '2023-9-21 11:00' & df$ts >='2023-9-14 16:30',NA,
                          ifelse(df$ch4_flux < -.20,NA,
                                 ifelse(df$ch4_flux > .3,NA,ch4_flux.limits)))
                                        
                                
hist(ch4_flux.limits,breaks = 100)
plot(ch4_flux.limits)

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(x=ts,y=ch4_flux),col = 'red')+
  geom_point(aes(x=ts,y=ch4_flux.limits),col = 'black')+
  scale_y_continuous(limits = c(-.35,.75))+
  scale_x_datetime(limits = as.POSIXct(c('2023-8-5','2023-10-12'),format="%F"))
```

### H

Plot H Histogram and what we might clean

```{r, warning=FALSE}
hist(df$H,breaks=100)

plot(df$ts,df$H,ylim = c(-250,450));abline(h=-150, col= 'red');abline(h=360, col='red')
```

Clean based on graphs, create new variable (H.limits), and see cleaned product

```{r, warning=FALSE}
H.limits = replace(df$H, df$H < -150 | df$H > 360,NA)
hist(H.limits,breaks = 100)
plot(H.limits)
```

### LE

Plot LE Histogram and what we might clean


```{r, warning=FALSE}
hist(df$LE,breaks=100)

plot(df$ts,df$LE,ylim = c(-40,450));abline(h=-40, col= 'red');abline(h=390, col='red')
```

Clean based on graphs, create new variable (H.limits), and see cleaned product

```{r, warning=FALSE}
LE.limits = replace(df$LE, df$LE < -40 | df$LE > 390,NA)
hist(LE.limits)
plot(LE.limits)
```

## USTAR

use night-time only flux values and plot Ustar based on data drop off

```{r, warning=FALSE}
df$nightCO2 = co2_flux.limits
df$nightCO2 = replace(df$nightCO2,df$SWIN_6_10_1_1_1 > 10,NA)

Uthreshold=0.065
ggplot(data = df,aes(`u*`,nightCO2))+
  ylab(expression(paste("Flux (g-C", ~CO[2], ~"m",""^"-2","d",""^"-1",")")))+
  geom_point()+xlab("U*")+
  scale_x_continuous(limits = c(0,1))+
  geom_vline(xintercept = Uthreshold,col='red')

```

pick threshold and filter based on plots above, I liked .065


created flag to indicate where u* values are below our threshold- makes cleaning with u* optional but easy when using cleaned data sets
```{r, warning=FALSE}
df$u_flag = ifelse(df$`u*` < Uthreshold,1,0)
```

## Combine Cleaning

investigate the previous limits and combine all to create a new variable (clean), plot to see difference

```{r, warning=FALSE}
co2fluxclean = ifelse(is.na(co2_flux.qc) | is.na(co2_flux.limits),NA,df$co2_flux)

ch4fluxclean = ifelse(is.na(ch4_flux.qc) | is.na(ch4_flux.limits),NA,df$ch4_flux)

Hclean = ifelse(is.na(H.qc) | is.na(H.limits),NA,df$H)

LEclean = ifelse(is.na(LE.qc) | is.na(LE.limits),NA,df$LE)

Tauclean = ifelse(is.na(Tau.qc) | is.na(Tau.limits),NA,df$Tau)

plot(df$ts,co2fluxclean)
plot(df$ts,ch4fluxclean)
plot(df$ts,Hclean)
plot(df$ts,LEclean)
plot(df$ts,Tauclean)
```

Add cleaned tau to dataframe (no MAD filter)
```{r, warning=FALSE}
df$Tau.c = Tauclean
```

# MAD Filter

Now, Spike filtering on half-hourly fluxes. For this type of data, filters based on median rather than std perform better. Use Unsymmetric Distributions and double MAD -MAD=Median Absolute Deviation for more info see link below <https://eurekastatistics.com/using-the-median-absolute-deviation-to-find-outliers/>

## NEE
```{r, warning=FALSE}
f = co2fluxclean # first make a new flux vector to apply the above filters
# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
Flux_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
Flux_filt = f                          # make a copy of the full fluxes again
Flux_filt[ind] = Flux_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = Flux_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(Flux_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(ts,co2fluxclean,col='CO2 flux'))+
  geom_line(aes(ts,Flux_filt,col='Filter'))+
  geom_line(aes(ts,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-10,10))+
  scale_x_datetime(limits = as.POSIXct(flux_dates,format="%F"))+
  scale_color_manual(values=c('black','yellow','red'))+
  geom_vline(xintercept = as.POSIXct('2024-01-01'))
```

now filter the data based on this signal

```{r, warning=FALSE}
#adjust this value (N) to tweak how many MADs to be away from the norm
N = 3

#Add clean value to dataframe (_.c)

df$co2_flux.c = ifelse(abs(co2fluxclean - Flux_filt) > (N * roll.mad),NA,co2fluxclean)

#Compare cleaned vs. uncleaned
ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,co2_flux),col='red')+
  geom_point(aes(TIMESTAMP,co2_flux.c),col='black')+
  scale_y_continuous(limits=c(-12,8))+
  scale_x_datetime(limits = as.POSIXct(flux_dates,format="%F"))
```

```{r, warning=FALSE}

ggplot(data = df)+theme_bw()+geom_hline(yintercept = 0)+
  geom_point(aes(ts,co2_flux), col='red')+
  geom_point(aes(ts,co2_flux.c), col='black')+
  scale_x_datetime(limits = as.POSIXct(c('2022-8-14','2022-8-28'),format="%F"))+
  scale_y_continuous (expression('NEE ('*mu*mol~m^-2~s^-1*')'),
                      limits = c(-20,20))

ggplot(data = df)+theme_bw()+geom_hline(yintercept = 0)+
  geom_point(aes(ts,co2_flux))+
  scale_x_datetime(limits = as.POSIXct(c('2022-9-24','2022-10-8'),format="%F"))+
  scale_y_continuous (expression('NEE ('*mu*mol~m^-2~s^-1*')'),
                      limits = c(-5,5))
```

See data with multiple layers of the MAD filter to see if it should be changed

```{r, warning=FALSE}
ggplot(data = df)+
  geom_ribbon(aes(x = ts,ymax = Flux_filt+roll.mad*6,ymin = Flux_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = ts,ymax = Flux_filt+roll.mad*3,ymin = Flux_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = ts,ymax = Flux_filt+roll.mad*2,ymin = Flux_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = ts,ymax = Flux_filt+roll.mad*1,ymin = Flux_filt-roll.mad*1,fill='1'))+
  geom_point(aes(ts,co2fluxclean),col='red')+
  geom_point(aes(ts,co2_flux.c),col='black')+
  scale_y_continuous(limits = c(-1,1))+
scale_x_datetime(limits = as.POSIXct(c('2023-2-3','2023-3-15'),format="%F"))
```

## CH4
```{r, warning=FALSE}
 #apply the moving window filter
f = ch4fluxclean # first make a new flux vector to apply the above filters

# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
Flux_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
Flux_filt = f                          # make a copy of the full fluxes again
Flux_filt[ind] = Flux_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = Flux_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(Flux_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(ts,ch4fluxclean,col='CH4 flux'))+
  geom_line(aes(ts,Flux_filt,col='Filter'))+
  geom_line(aes(ts,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-.25,.3))+
  scale_x_datetime(limits = as.POSIXct(flux_dates,format="%F"))+
  scale_color_manual(values=c('black','yellow','red'))
```

adjust this (N) to tweak how many MADs to be away from the norm, CH4 deserves more because of it's nature, add clean value to the dataframe (_.c) and compare cleaned vs. uncleaned


```{r, warning=FALSE}
N = 6 

df$ch4_flux.c = ifelse(abs(ch4fluxclean - Flux_filt) > (N * roll.mad),NA,ch4fluxclean)

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,ch4_flux,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,ch4_flux.c,col='cleaned'))+
  scale_y_continuous(limits = c(-.25,.3))+
  scale_x_datetime(limits = as.POSIXct(flux_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))
```

See data with multiple layers of the MAD filter to see if it should be changed

```{r, warning=FALSE}
CH4Filter = ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(df$ts,ch4fluxclean),col='red')+
  geom_point(aes(df$ts,ch4_flux.c),col='black')+
  scale_y_continuous(limits = c(-.22,.25))+
  scale_x_datetime(limits = as.POSIXct(c('2023-9-1','2023-11-2'),format="%F"))

CH4RSSI = ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_line(aes(x=ts,y=rssi_77_mean))+
  scale_y_continuous(limits = c(-1,78))+
  scale_x_datetime(limits = as.POSIXct(c('2023-9-1','2023-11-2'),format="%F"))


CH4MOLE = ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(x=ts,y=ch4_mole_fraction))+
  scale_y_continuous(limits = c(0,4))+
  scale_x_datetime(limits = as.POSIXct(c('2023-9-1','2023-11-2'),format="%F"))
CH4MOLE

CH4_FALLSOILTCOMP = plot_grid(CH4Filter,CH4RSSI,CH4MOLE,nrow=3,align = 'v')
CH4_FALLSOILTCOMP

```


## H
```{r, warning=FALSE}
# apply the moving window filter
f = Hclean # first make a new flux vector to apply the above filters

# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
Flux_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
Flux_filt = f                          # make a copy of the full fluxes again
Flux_filt[ind] = Flux_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = Flux_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(Flux_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(ts,Hclean,col='H'))+
  geom_line(aes(ts,Flux_filt,col='Filter'))+
  geom_line(aes(ts,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-180,350))+
  scale_x_datetime(limits = as.POSIXct(flux_dates,format="%F"))+
  scale_color_manual(values=c('yellow','black','red'))
```

Now filter the data based on this signal (N) adjust this to tweak how many MADs to be away from the norm, add clean value to dataframe (_.c) and Compare Clean vs. Uncleaned

```{r, warning=FALSE}
N = 3

df$H.c = ifelse(abs(Hclean - Flux_filt) > (N * roll.mad),NA,Hclean)

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,Hclean,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,H.c,col='cleaned'))+
  scale_y_continuous(limits=c(-150,350))+
  scale_x_datetime(limits = as.POSIXct(flux_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))
```

See data with multiple layers of the MAD filter to see if it should be changed

```{r, warning=FALSE}
ggplot(data = df)+
  geom_ribbon(aes(x = df$ts,ymax = Flux_filt+roll.mad*6,ymin = Flux_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = df$ts,ymax = Flux_filt+roll.mad*3,ymin = Flux_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = df$ts,ymax = Flux_filt+roll.mad*2,ymin = Flux_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = df$ts,ymax = Flux_filt+roll.mad*1,ymin = Flux_filt-roll.mad*1,fill='1'))+
  geom_point(aes(df$ts,Hclean),col='red')+
  geom_point(aes(df$ts,H.c),col='black')+
  scale_y_continuous(limits = c(-100,100))+
  scale_x_datetime(limits = as.POSIXct(c('2023-1-19','2023-2-24')))
```

## LE

```{r, warning=FALSE}
# apply the moving window filter
f = LEclean # first make a new flux vector to apply the above filters

# prepare the inputs for a zero-phase (forward and reverse) filter  
ind       = which(complete.cases(f))   # first save the index of all the NAs
b         = rep(1,8)/8                 # moving average coefficient
a         = 1                          # coefficient of the ARMA filter
Flux_i    = filtfilt(b, a, na.omit(f)) #create a forward and reverse ARMA filter
Flux_filt = f                          # make a copy of the full fluxes again
Flux_filt[ind] = Flux_i                # apply the filter where the real data exist

n = 7*48 # set the length of the rolling filter, 48 half hours in day * x days
roll.mad = rollapply(data = Flux_filt,
                     width = n,
                     FUN = mad,
                     partial = TRUE,
                     na.rm = T,
                     fill = NA)  #create a rolling MAD or sd or other
roll.mad[is.na(Flux_filt)] = NA  #make NA values where NAs exist in the real data

ggplot(df)+geom_hline(yintercept = 0)+
  geom_point(aes(ts,LEclean,col='LE'))+
  geom_line(aes(ts,Flux_filt,col='Filter'))+
  geom_line(aes(ts,roll.mad,col='MAD'))+
  scale_y_continuous(limits = c(-50,400))+
  scale_x_datetime(limits = as.POSIXct(flux_dates,format="%F"))+
  scale_color_manual(values=c('yellow','black','red'))
```

Now filter the data based on this signal (N) adjust this to tweak how many MADs to be away from the norm

```{r, warning=FALSE}
N = 6

df$LE.c = ifelse(abs(LEclean - Flux_filt) > (N * roll.mad),NA,LEclean)

ggplot(data = df)+geom_hline(yintercept = 0)+
  geom_point(aes(TIMESTAMP,LEclean,col='uncleaned'))+
  geom_point(aes(TIMESTAMP,LE.c,col='cleaned'))+
  scale_y_continuous(limits=c(-100,400))+
  scale_x_datetime(limits = as.POSIXct(flux_dates,format="%F"))+
  scale_color_manual(values=c('black','red'))
```

See data with multiple layers of the MAD filter to see if it should be changed

```{r, warning=FALSE}
ggplot(data = df)+
  geom_ribbon(aes(x = df$ts,ymax = Flux_filt+roll.mad*6,ymin = Flux_filt-roll.mad*6,fill='6'))+
  geom_ribbon(aes(x = df$ts,ymax = Flux_filt+roll.mad*3,ymin = Flux_filt-roll.mad*3,fill='3'))+
  geom_ribbon(aes(x = df$ts,ymax = Flux_filt+roll.mad*2,ymin = Flux_filt-roll.mad*2,fill='2'))+
  geom_ribbon(aes(x = df$ts,ymax = Flux_filt+roll.mad*1,ymin = Flux_filt-roll.mad*1,fill='1'))+
  geom_point(aes(df$ts,LEclean),col='red')+
  geom_point(aes(df$ts,LE.c),col='black')+
  scale_y_continuous(limits = c(-20,20))+
  scale_x_datetime(limits = as.POSIXct(c('2022-11-20','2023-2-15')))
```

# Main Plots

```{r, warning=FALSE}
ggplot(data = df)+theme_bw()+geom_hline(yintercept = 0)+
  geom_point(aes(ts,LE,col='Filt.'))+
  geom_point(aes(ts,LE.c,col='clean'))+
  scale_y_continuous(expression('LE ('*W~m^-2*')'))+
  scale_color_manual(values = c('black','red'))
  

ggplot(data = df)+theme_bw()+geom_hline(yintercept = 0)+
  geom_point(aes(ts,H,col='Filt.'))+
  geom_point(aes(ts,H.c,col='clean'))+
  scale_y_continuous(expression('H ('*W~m^-2*')'))+
  scale_color_manual(values = c('black','red'))

ggplot(data = df)+theme_bw()+geom_hline(yintercept = 0)+
  geom_point(aes(ts,co2_flux,col='Filt.'))+
  geom_point(aes(ts,co2_flux.c,col='clean'))+
  scale_y_continuous (expression('NEE ('*mu*mol~m^-2~s^-1*')'),
                      limits = c(-20,20))+
  scale_color_manual(values = c('black','red'))

ggplot(data = df)+theme_bw()+geom_hline(yintercept = 0)+
  geom_point(aes(ts,ch4_flux*43.2,col='Filt.'))+
  geom_point(aes(ts,ch4_flux.c*43.2,col='clean'))+
  scale_y_continuous(expression(CH[4]~'flux ('*mu*mol~m^-2~s^-1*')'),
                     limits = c(-3,8))+
  scale_color_manual(values = c('black','red'))

```

# Diurnal Fluxes

set hour and month vectors to be able to take the averages

```{r, warning=FALSE}
df$hour = as.numeric(format(df$ts,'%H'))
df$month = as.numeric(format(df$ts,'%m'))

diel = df %>%
  group_by(hour,month) %>%
  select(co2_flux.c,ch4_flux.c,LE.c,H.c,hour,month) %>%
  summarise_all(.funs = c(median,sd),na.rm=T)

diel = subset(diel,!is.na(diel$month))
```

## NEE

```{r, warning=FALSE}
ggplot(data = diel)+geom_hline(yintercept = 0)+
  geom_line(aes(hour,co2_flux.c_fn1),lty=2)+
  geom_ribbon(aes(x = hour,y = co2_flux.c_fn1,
                  ymax = co2_flux.c_fn1 + co2_flux.c_fn2,
                  ymin = co2_flux.c_fn1 - co2_flux.c_fn2),
              alpha= 0.5)+
  facet_wrap(~month)
levels(diel$month)
```

## CH4

```{r, warning=FALSE}
ggplot(data = diel)+geom_hline(yintercept = 0)+
  geom_line(aes(hour,ch4_flux.c_fn1),lty=2)+
  geom_ribbon(aes(x = hour,y = ch4_flux.c_fn1,
                  ymax = ch4_flux.c_fn1 + ch4_flux.c_fn2,
                  ymin = ch4_flux.c_fn1 - ch4_flux.c_fn2),
              alpha= 0.5)+
  facet_wrap(~month)

ggplot(data = diel)+geom_hline(yintercept = 0)+
  geom_line(aes(hour,ch4_flux.c_fn1*43.2),lty=2)+
  geom_ribbon(aes(x = hour,y = ch4_flux.c_fn1*43.2,
                  ymax = ch4_flux.c_fn1*43.2 + ch4_flux.c_fn2*43.2,
                  ymin = ch4_flux.c_fn1*43.2 - ch4_flux.c_fn2*43.2),
              alpha= 0.5)+
  scale_y_continuous(limits = c(-1,6))+
  facet_wrap(~month)
```

## H

```{r, warning=FALSE}
ggplot(data = diel)+geom_hline(yintercept = 0)+
  geom_line(aes(hour,H.c_fn1),lty=2)+
  geom_ribbon(aes(x = hour,y = H.c_fn1,
                  ymax = H.c_fn1 + H.c_fn2,
                  ymin = H.c_fn1 - H.c_fn2),
              alpha= 0.5)+
  facet_wrap(~month)
```

## LE

```{r, warning=FALSE}
ggplot(data = diel)+geom_hline(yintercept = 0)+
  geom_line(aes(hour,LE.c_fn1),lty=2)+
  geom_ribbon(aes(x = hour,y = LE.c_fn1,
                  ymax = LE.c_fn1 + LE.c_fn2,
                  ymin = LE.c_fn1 - LE.c_fn2),
              alpha= 0.5)+
  facet_wrap(~month)
```

# Save

```{r, warning=FALSE}
df <- df[!duplicated(df$TIMESTAMP), ]

write.csv(x = df,file = 'C:/Users/dtrangmoe/Documents/Churchill/Churchill_Flux_Met_Clean_Ameriflux_2022_2023.csv',row.names = F,quote = F)
```
